大三2/人工神经网络/project/src/requirements.txt
h5py==3.10.0
matplotlib==3.7.1
numpy==1.24.2
opencv_python==4.8.0.74
pandas==2.0.3
Pillow==10.2.0
Pillow==10.2.0
scikit_learn==1.3.0
tensorboardX==2.6.2.2
torchsummary
tqdm==4.61.2
transformers==4.38.2
